requirements/dev.txt
-r base.txt

# django development
django-debug-toolbar==1.11
django-extensions==2.1.6

# tests
selenium==3.5.0
pytest==4.4.1
pytest-django==3.4.8

# docs
watchdog==0.8.3
Sphinx==1.6.3

# django_extensions.graph_models
pydot==1.2.3
